pyproject.toml
[project]
name = "langgraph-complex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-openai>=0.3.16",
    "langchain[anthropic]>=0.3.25",
    "langgraph>=0.4.3",
    "python-dotenv>=1.1.0",
]
